app-ops/requirements.txt
boto3==1.26.133
python-dotenv
twilio==8.1.0
